requirements.txt
gradio==4.44.0
llama-cpp-python==0.3.2
huggingface-hub>=0.20.0